[FLINK-1297] Added an accumulator called OperatorStatsAccumulator capable of tracking min, max and estimates for count distinct and heavy hitters.
The count distinct algorithms are Linear Counting and HyperLogLog, both from an imported library from clearspring.

The heavy hitters algorithms are Lossy counting (Manku et.al 2002) and one based on Count Min Sketch (Cormode 2005).

The heavy hitters algorithms are implemented in the statistics package in flink-contrib.

Refactored to flink-contrib.
Handled exceptions properly.
Renamed IHeavyHitter

Changed the default algorithm for Count Distinct to be HyperLogLog since the class for LinearCounting is not serializable and this was causing an error. Made LossyCounting class serializable as well.

The OperatorStatsAccumulatorTest creates an accumulator for each subtask, in order to track local as well as global stats.